Changed the position of the "Limpar" button and finished cleaning the code
