+++ src/components/layout/SidebarPosition.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useState, useEffect } from "react";
+import { createContext, useContext, useState } from "react";
 
 type SidebarContextType = {
     sidebarPosition: 'left' | 'right';
