+++ src/pages/TimeTableSelector.tsx
@@ -6,6 +6,7 @@ import { CourseInfo, Major } from '../@types'
 import MajorContext from '../contexts/MajorContext'
 import CourseContext from '../contexts/CourseContext'
 import { useSidebarContext } from '../components/layout/SidebarPosition'
+import { SidebarProvider } from '../components/layout/SidebarPosition'
 
 const TimeTableSelectorPage = () => {
   const [majors, setMajors] = useState<Major[]>([])
@@ -13,7 +14,6 @@ const TimeTableSelectorPage = () => {
   const [pickedCourses, setPickedCourses] = useState<CourseInfo[]>(StorageAPI.getPickedCoursesStorage());
   const [checkboxedCourses, setCheckboxedCourses] = useState<CourseInfo[]>(StorageAPI.getPickedCoursesStorage());
   const [ucsModalOpen, setUcsModalOpen] = useState<boolean>(false);
-  const { sidebarPosition } = useSidebarContext();
 
   //TODO: Looks suspicious
   const [choosingNewCourse, setChoosingNewCourse] = useState<boolean>(false);
@@ -36,31 +36,41 @@ const TimeTableSelectorPage = () => {
           ucsModalOpen, setUcsModalOpen
         }
       }>
-        <div className="grid w-full grid-cols-12 gap-x-4 gap-y-4 px-4 py-4">
-          {sidebarPosition === 'left' ?(
-            <>
-              <div className='col-span-12 lg:col-span-3 min-h'>
-                <Sidebar />
-              </div>
-              <div className='col-span-12 lg:col-span-9 min-h rounded-md bg-lightest px-3 py-3 dark:bg-dark 2xl:px-5 2xl:py-5'>
-                <Schedule />
-              </div>
-            </>
-          ) : (
-            <>
-              <div className='col-span-12 lg:col-span-9 min-h rounded-md bg-lightest px-3 py-3 dark:bg-dark 2xl:px-5 2xl:py-5'>
-              <Schedule />
-              </div>
-              <div className='col-span-12 lg:col-span-3 min-h'>
-              <Sidebar />
-              </div>
-            </>
-          )}
-        </div>
+        <SidebarProvider>
+          <Content />
+        </SidebarProvider>
       </CourseContext.Provider>
     </MajorContext.Provider>
   );
 };
+
+const Content = () => {
+  const { sidebarPosition } = useSidebarContext();
+
+  return (
+    <div className="grid w-full grid-cols-12 gap-x-4 gap-y-4 px-4 py-4">
+      {sidebarPosition === 'left' ?(
+        <>
+          <div className='col-span-12 lg:col-span-3 min-h'>
+            <Sidebar />
+          </div>
+          <div className='col-span-12 lg:col-span-9 min-h rounded-md bg-lightest px-3 py-3 dark:bg-dark 2xl:px-5 2xl:py-5'>
+            <Schedule />
+          </div>
+        </>
+      ) : (
+        <>
+          <div className='col-span-12 lg:col-span-9 min-h rounded-md bg-lightest px-3 py-3 dark:bg-dark 2xl:px-5 2xl:py-5'>
+          <Schedule />
+          </div>
+          <div className='col-span-12 lg:col-span-3 min-h'>
+          <Sidebar />
+          </div>
+        </>
+      )}
+    </div>
+  )
+}
 
 
 export default TimeTableSelectorPage;
